Simplified scale offsets

+++ src/lib/quantizer/Note.h
@@ -4,11 +4,11 @@
 class Note {
     public:
         Note() {}
-        Note(int cycle, int note, float voltage = 0, int offset = 0) 
+        Note(int cycle, int note, float voltage = 0, float offset = 0) 
             { this->cycle = cycle; this->note = note; this->voltage = voltage; this->offset = offset; }
         int cycle = 0;
         int note = 0;
-        int offset = 0;
+        float offset = 0;
         float voltage = 0;
 };
